Refactor explainer_rule_eval into submodules
