+++ tests/test_categorical_optimizer.py
@@ -242,7 +242,7 @@ def test_cli_global_steps_multiple_generations(tmp_path, monkeypatch):
     def fake_eval(*a, optimizer=None, **k):
         trial = {"detected": True, "condition_type": "any", "combine_mode": "or"}
         if optimizer is not None:
-            from src.cli.explainer_rule_eval import _optimizer_step
+            from src.cli.explainer_rule_eval.optimizer_utils import _optimizer_step
             _optimizer_step(trial, optimizer)
         return trial
