+++ tests/test_fp_exclude_persist.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 from concurrent.futures import ProcessPoolExecutor
 
-from src.cli.explainer_rule_eval import _save_fp_exclude, _load_fp_exclude
+from src.cli.explainer_rule_eval.fp_token_utils import _save_fp_exclude, _load_fp_exclude
 
 
 def test_save_and_load_fp_exclude(tmp_path):
